requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
# 火山 Ark 模型（可选，未安装时对话使用关键词意图识别）
volcengine-python-sdk[ark]
# MCP 支持（可选，用于工具调用）
fastmcp>=0.1.0
